requirements.txt
google-genai>=1.0.0
python-docx>=1.1.0
odfpy>=1.4.1
docx2pdf>=0.1.8

